+++ public/cpp/logic/public/public/gm/actor_gmclient.cpp
@@ -56,7 +56,7 @@ namespace ngl
 			);
 		}
 
-		if (handle_cmd::function(loperator, adata.get_data()->identifier(), lojson) == false)
+		if (handle_cmd::function(loperator, (int32_t)adata.get_data()->identifier(), lojson) == false)
 		{
 			log_error()->print("GM actor_gmclient operator[{}] ERROR", loperator);
 		}
